Finalize helper functions for 0.6 (#2632)
