+++ beacon-chain/core/helpers/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
         "randao.go",
         "rewards_penalties.go",
         "slot_epoch.go",
-        "state.go",
         "validators.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/helpers",
@@ -38,7 +37,6 @@ go_test(
         "randao_test.go",
         "rewards_penalties_test.go",
         "slot_epoch_test.go",
-        "state_test.go",
         "validators_test.go",
     ],
     embed = [":go_default_library"],
