+++ beacon-chain/core/helpers/validators.go
@@ -153,3 +153,30 @@ func BeaconProposerIndex(state *pb.BeaconState) (uint64, error) {
 		}
 	}
 }
+
+// DomainVersion returns the domain version for BLS private key to sign and verify.
+//
+//  def get_domain(state: BeaconState,
+//               domain_type: int,
+//               message_epoch: int=None) -> int:
+//    """
+//    Return the signature domain (fork version concatenated with domain type) of a message.
+//    """
+//    epoch = get_current_epoch(state) if message_epoch is None else message_epoch
+//    fork_version = state.fork.previous_version if epoch < state.fork.epoch else state.fork.current_version
+//    return bytes_to_int(fork_version + int_to_bytes(domain_type, length=4))
+func DomainVersion(state *pb.BeaconState, epoch uint64, domainType uint64) uint64 {
+	if epoch == 0 {
+		epoch = CurrentEpoch(state)
+	}
+	var forkVersion []byte
+	if epoch < state.Fork.Epoch {
+		forkVersion = state.Fork.PreviousVersion
+	} else {
+		forkVersion = state.Fork.CurrentVersion
+	}
+	by := []byte{}
+	by = append(by, forkVersion[:4]...)
+	by = append(by, bytesutil.Bytes4(domainType)...)
+	return bytesutil.FromBytes8(by)
+}
